refactor(dh): move emDash variable to shared folder
