+++ libs/dh/metering-point/feature-overview/src/lib/primary-master-data/dh-metering-point-primary-master-data.component.spec.ts
@@ -31,7 +31,7 @@ import {
   DhMeteringPointPrimaryMasterDataScam,
   PrimaryMasterData,
 } from './dh-metering-point-primary-master-data.component';
-import { emDash } from '../identity/em-dash';
+import { emDash } from '../shared/em-dash';
 
 describe(DhMeteringPointPrimaryMasterDataComponent.name, () => {
   const fallbackValue = emDash;
