+++ libs/dh/metering-point/feature-overview/src/lib/primary-master-data/dh-metering-point-primary-master-data.component.ts
@@ -24,6 +24,7 @@ import {
   SimpleChanges,
 } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { DomSanitizer } from '@angular/platform-browser';
 import { TranslocoModule } from '@ngneat/transloco';
 
 import { MeteringPointCimDto } from '@energinet-datahub/dh/shared/data-access-api';
@@ -32,9 +33,9 @@ import {
   WattIconModule,
   WattIconSize,
 } from '@energinet-datahub/watt';
-import { DomSanitizer } from '@angular/platform-browser';
-import { emDash } from '../identity/em-dash';
+
 import { DhEmptyValuePipeScam } from './empty-value.pipe';
+import { emDash } from '../shared/em-dash';
 import { DhYesNoPipeScam } from '../shared/yes-no.pipe';
 
 export type PrimaryMasterData = Pick<
